Start extraction of common TDP logic.
