+++ src/tree/v2/ResourceTreeDataProviderBase.ts
@@ -1,0 +1,90 @@
+import * as vscode from 'vscode';
+import { ApplicationResourceProviderManager } from '../../api/v2/providers/ApplicationResourceProviderManager';
+import { ResourceModelBase } from '../../api/v2/v2AzureResourcesApi';
+import { BranchDataProviderManager } from './providers/BranchDataProviderManager';
+import { ResourceGroupsItem } from './ResourceGroupsItem';
+import { ResourceGroupsItemCache } from './ResourceGroupsItemCache';
+
+export abstract class ResourceTreeDataProviderBase extends vscode.Disposable implements vscode.TreeDataProvider<ResourceGroupsItem> {
+    private readonly branchChangeSubscription: vscode.Disposable;
+    private readonly refreshSubscription: vscode.Disposable;
+    private readonly resourceProviderManagerListener: vscode.Disposable;
+    protected readonly onDidChangeTreeDataEmitter = new vscode.EventEmitter<void | ResourceGroupsItem | ResourceGroupsItem[] | null | undefined>();
+
+    constructor(
+        branchDataProviderManager: BranchDataProviderManager,
+        protected readonly itemCache: ResourceGroupsItemCache,
+        refreshEvent: vscode.Event<void>,
+        protected readonly resourceProviderManager: ApplicationResourceProviderManager,
+        callOnDispose: () => void) {
+        super(
+            () => {
+                callOnDispose();
+
+                this.branchChangeSubscription.dispose();
+                this.refreshSubscription.dispose();
+                this.resourceProviderManagerListener.dispose();
+            });
+
+        this.branchChangeSubscription = branchDataProviderManager.onDidChangeTreeData(
+            (e: void | ResourceModelBase | ResourceModelBase[] | null | undefined) => {
+                const rgItems: ResourceGroupsItem[] = [];
+
+                // eslint-disable-next-line no-extra-boolean-cast
+                if (!!e) {
+                    // e was defined, either a single item or array
+                    // Make an array for consistency
+                    const branchItems: unknown[] = Array.isArray(e) ? e : [e];
+
+                    for (const branchItem of branchItems) {
+                        const rgItem = this.itemCache.getItemForBranchItem(branchItem);
+
+                        if (rgItem) {
+                            rgItems.push(rgItem);
+                        }
+                    }
+                } else {
+                    // e was null/undefined/void
+                    // Translate it to fire on all elements for this branch data provider
+                    // TODO
+                }
+
+                this.onDidChangeTreeDataEmitter.fire(rgItems)
+            });
+
+        this.refreshSubscription = refreshEvent(() => this.onDidChangeTreeDataEmitter.fire());
+
+        // TODO: If only individual resources change, just update the tree related to those resources.
+        this.resourceProviderManagerListener = resourceProviderManager.onDidChangeResourceChange(() => this.onDidChangeTreeDataEmitter.fire());
+    }
+
+    onDidChangeTreeData: vscode.Event<void | ResourceGroupsItem | ResourceGroupsItem[] | null | undefined> = this.onDidChangeTreeDataEmitter.event;
+
+    getTreeItem(element: ResourceGroupsItem): vscode.TreeItem | Thenable<vscode.TreeItem> {
+        return element.getTreeItem();
+    }
+
+    async getChildren(element?: ResourceGroupsItem | undefined): Promise<ResourceGroupsItem[] | null | undefined> {
+        if (element) {
+            // TODO: Do we really need to evict before generating new children, or can we just update after the fact?
+            //       Since the callback is async, could change notifications show up while doing this?
+            this.itemCache.evictItemChildren(element);
+        } else {
+            this.itemCache.evictAll();
+        }
+
+        const children = await this.onGetChildren(element);
+
+        if (children) {
+            if (element) {
+                this.itemCache.updateItemChildren(element, children);
+            } else {
+                children.forEach(child => this.itemCache.addItem(child, []));
+            }
+        }
+
+        return children;
+    }
+
+    protected abstract onGetChildren(element?: ResourceGroupsItem | undefined): Promise<ResourceGroupsItem[] | null | undefined>;
+}
